RiddleProcessor/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.6.2.post1=pypi_0
attrs=24.2.0=pypi_0
av=12.3.0=pypi_0
bidict=0.23.1=pypi_0
blas=1.0=mkl
blinker=1.8.2=pypi_0
brotli-python=1.0.9=py311hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.9.24=haa95532_0
certifi=2024.8.30=py311haa95532_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
ctranslate2=4.5.0=pypi_0
cuda-cccl=12.6.77=0
cuda-cccl_win-64=12.6.77=0
cuda-cudart=12.4.127=0
cuda-cudart-dev=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-libraries-dev=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvrtc-dev=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.77=0
cuda-opencl-dev=12.6.77=0
cuda-profiler-api=12.6.77=0
cuda-runtime=12.4.1=0
cuda-version=12.6=3
distro=1.9.0=pypi_0
dnspython=2.7.0=pypi_0
eventlet=0.37.0=pypi_0
faster-whisper=1.0.3=pypi_0
filelock=3.16.1=pypi_0
flask=3.0.3=pypi_0
flask-socketio=5.4.1=pypi_0
flatbuffers=24.3.25=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
giflib=5.2.2=h7edc060_0
gmpy2=2.1.2=py311h7f96b67_0
greenlet=3.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.26.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
inflection=0.5.1=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=py311haa95532_0
jiter=0.6.1=pypi_0
jpeg=9e=h827c3e9_3
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libcublas=12.4.5.8=0
libcublas-dev=12.4.5.8=0
libcufft=11.2.1.3=0
libcufft-dev=11.2.1.3=0
libcurand=10.3.7.77=0
libcurand-dev=10.3.7.77=0
libcusolver=11.6.1.9=0
libcusolver-dev=11.6.1.9=0
libcusparse=12.3.1.170=0
libcusparse-dev=12.3.1.170=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.2.5.30=0
libnpp-dev=12.2.5.30=0
libnvfatbin=12.6.77=0
libnvfatbin-dev=12.6.77=0
libnvjitlink=12.4.127=0
libnvjitlink-dev=12.4.127=0
libnvjpeg=12.3.1.117=0
libnvjpeg-dev=12.3.1.117=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
llvmlite=0.43.0=pypi_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=3.0.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.10=py311h827c3e9_0
mkl_random=1.2.7=py311hea22821_0
more-itertools=10.5.0=pypi_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
multidict=6.1.0=pypi_0
networkx=3.4.1=pypi_0
numba=0.60.0=pypi_0
numpy=2.0.2=pypi_0
numpy-base=2.0.1=py311hd01c5d8_1
onnxruntime=1.19.2=pypi_0
openai=1.52.0=pypi_0
openai-whisper=20240930=pypi_0
openapi=2.0.0=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.15=h827c3e9_0
packaging=24.1=pypi_0
pillow=10.4.0=py311h827c3e9_0
pip=24.2=py311haa95532_0
propcache=0.2.0=pypi_0
protobuf=5.28.3=pypi_0
pyaudio=0.2.14=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydub=0.25.1=pypi_0
pyreadline3=3.5.4=pypi_0
pysocks=1.7.1=py311haa95532_0
python=3.11.10=h4607a30_0
python-dotenv=1.0.1=pypi_0
python-engineio=4.10.1=pypi_0
python-socketio=5.11.4=pypi_0
pytorch=2.5.0=py3.11_cuda12.4_cudnn9_0
pytorch-cuda=12.4=h3fd98bf_7
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py311h827c3e9_0
referencing=0.35.1=pypi_0
regex=2024.9.11=pypi_0
requests=2.32.3=py311haa95532_0
rpds-py=0.20.0=pypi_0
setuptools=75.1.0=py311haa95532_0
simple-websocket=1.1.0=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
tiktoken=0.8.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.20.1=pypi_0
torchaudio=2.5.0=pypi_0
torchvision=0.20.0=pypi_0
tqdm=4.66.5=pypi_0
typing-extensions=4.12.2=pypi_0
typing_extensions=4.11.0=py311haa95532_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py311haa95532_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
websocket-client=1.8.0=py311haa95532_0
werkzeug=3.0.4=pypi_0
wheel=0.44.0=py311haa95532_0
whisper=1.1.10=pypi_0
win_inet_pton=1.1.0=py311haa95532_0
wsproto=1.2.0=pypi_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.15.5=pypi_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
